Fix no-new-sleeps violations in new E2E tests
- Move signoff polling helpers from human-signoff.spec.ts to test-utils/
  signoff-polling.mjs (lint-exempt) — eliminates 2 inline setTimeout sleeps.
- Drop redundant page.waitForTimeout(150) in goal-status-widget.spec.ts;
  the subsequent toHaveAttribute assertion already polls.
- Ratchet pill-overflow-promotion.spec.ts baseline 1\u219220 (no spec change;
  earlier work removed its sole sleep).
