+++ tests/e2e/test-utils/no-new-sleeps.baseline.json
@@ -7,6 +7,5 @@
 	"multi-repo-project.spec.ts": 1,
 	"pool-flow.spec.ts": 4,
 	"thinking-level.spec.ts": 1,
-	"ui/pill-overflow-promotion.spec.ts": 1,
 	"ui/review-annotations-persistence.spec.ts": 1
 }
